+++ extensions/bhagavad-gita-quotes/.eslintrc.cjs
@@ -3,8 +3,7 @@ Copyright © 2026 Sam Analytic Solutions
 All rights reserved.
 */
 
-import { defineConfig } from "eslint/config";
-import raycastConfig from "@raycast/eslint-config";
-
-export default defineConfig([...raycastConfig]);
-
+module.exports = {
+  root: true,
+  extends: ["@raycast/eslint-config"],
+};